op-deployer: remove legacy DeployAsterisc script
