+++ op-deployer/pkg/deployer/opcm/asterisc.go
@@ -1,30 +1,21 @@
 package opcm
 
 import (
-	"github.com/ethereum/go-ethereum/common"
-
 	"github.com/ethereum-optimism/optimism/op-chain-ops/script"
+	"github.com/ethereum/go-ethereum/common"
 )
 
 type DeployAsteriscInput struct {
 	PreimageOracle common.Address
 }
 
-func (input *DeployAsteriscInput) InputSet() bool {
-	return true
-}
-
 type DeployAsteriscOutput struct {
 	AsteriscSingleton common.Address
 }
 
-func (output *DeployAsteriscOutput) CheckOutput(input common.Address) error {
-	return nil
-}
+type DeployAsteriscScript script.DeployScriptWithOutput[DeployAsteriscInput, DeployAsteriscOutput]
 
-func DeployAsterisc(
-	host *script.Host,
-	input DeployAsteriscInput,
-) (DeployAsteriscOutput, error) {
-	return RunScriptSingle[DeployAsteriscInput, DeployAsteriscOutput](host, input, "DeployAsterisc.s.sol", "DeployAsterisc")
+// NewDeployAsteriscScript loads and validates the DeployAsterisc script contract
+func NewDeployAsteriscScript(host *script.Host) (DeployAsteriscScript, error) {
+	return script.NewDeployScriptWithOutputFromFile[DeployAsteriscInput, DeployAsteriscOutput](host, "DeployAsterisc.s.sol", "DeployAsterisc")
 }
