--- op-deployer/pkg/deployer/opcm/asterisc2_test.go
@@ -1,50 +1,0 @@
-package opcm
-
-import (
-	"math/big"
-	"testing"
-
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/stretchr/testify/require"
-)
-
-func TestNewDeployAsteriscScript(t *testing.T) {
-	t.Run("should not fail with current version of DeployAsterisc2 contract", func(t *testing.T) {
-		// First we grab a test host
-		host1 := createTestHost(t)
-
-		// Then we load the script
-		//
-		// This would raise an error if the Go types didn't match the ABI
-		deploySuperchain, err := NewDeployAsteriscScript(host1)
-		require.NoError(t, err)
-
-		// Then we deploy
-		output, err := deploySuperchain.Run(DeployAsterisc2Input{
-			PreimageOracle: common.BigToAddress(big.NewInt(1)),
-		})
-
-		// And do some simple asserts
-		require.NoError(t, err)
-		require.NotNil(t, output)
-
-		// Now we run the old deployer
-		//
-		// We run it on a fresh host so that the deployer nonces are the same
-		// which in turn means we should get identical output
-		host2 := createTestHost(t)
-		deprecatedOutput, err := DeployAsterisc(host2, DeployAsteriscInput{
-			PreimageOracle: common.BigToAddress(big.NewInt(1)),
-		})
-
-		// Make sure it succeeded
-		require.NoError(t, err)
-		require.NotNil(t, deprecatedOutput)
-
-		// Now make sure the addresses are the same
-		require.Equal(t, deprecatedOutput.AsteriscSingleton, output.AsteriscSingleton)
-
-		// And just to be super sure we also compare the code deployed to the addresses
-		require.Equal(t, host2.GetCode(deprecatedOutput.AsteriscSingleton), host1.GetCode(output.AsteriscSingleton))
-	})
-}
